requirements.txt
grpcio==1.51.3
pygnmi==0.8.15
PyYAML==6.0.2
